Move Search within Europeana::Record
And rename `Europeana::Record.get` to `.fetch`
