+++ lib/europeana/record/search/fields.rb
